// src/app/model/recordings.ts
export interface Recordings {
    conversationId:number;
    ani:string;
    dnis: string;
    cola: string;
    usuario: string;
    identificacion:string;
    fecha:string;
}

/*
conversationId:number;
    ani:string;
    dnis: string;
    cola: string;
    usuario: string;
    identificacion:string;
    fecha:string;


     userId:number;
    id:number;
    title: string;
    body: string;
*/ 